requirements.txt
flask
jinja2
python-dotenv
flask-sqlalchemy
requests
pytz
pillow
sqlalchemy
